♻️ [Refactor] 비교 페이지 렌더 로직 분리
프레임 및 변경사항 Fabric.js 객체 생성 비즈니스 로직 분리
